Refactor generic rule generator
