+++ packages/style-engine/src/styles/color/index.ts
@@ -1,6 +1,14 @@
 /**
  * Internal dependencies
  */
-import text from './text';
+import type { Style, StyleOptions } from '../../types';
+import { generateRule } from '../utils';
+
+const text = {
+	name: 'text',
+	generate: ( style: Style, options: StyleOptions ) => {
+		return generateRule( style, [ 'color', 'text' ], 'color', options );
+	},
+};
 
 export default [ text ];
