--- packages/style-engine/src/styles/typography/fontSize.ts
@@ -1,32 +1,0 @@
-/**
- * External dependencies
- */
-import { get } from 'lodash';
-
-/**
- * Internal dependencies
- */
-import type { Style, StyleOptions } from '../../types';
-
-const fontSize = {
-	name: 'fontSize',
-	generate: ( style: Style, options: StyleOptions ) => {
-		const styleValue: string | undefined = get(
-			style,
-			[ 'typography', 'fontSize' ],
-			null
-		);
-
-		return styleValue
-			? [
-					{
-						selector: options?.selector,
-						key: 'fontSize',
-						value: styleValue,
-					},
-			  ]
-			: [];
-	},
-};
-
-export default fontSize;
