--- packages/style-engine/src/styles/typography/textDecoration.ts
@@ -1,32 +1,0 @@
-/**
- * External dependencies
- */
-import { get } from 'lodash';
-
-/**
- * Internal dependencies
- */
-import type { Style, StyleOptions } from '../../types';
-
-const textDecoration = {
-	name: 'textDecoration',
-	generate: ( style: Style, options: StyleOptions ) => {
-		const styleValue: string | undefined = get(
-			style,
-			[ 'typography', 'textDecoration' ],
-			null
-		);
-
-		return styleValue
-			? [
-					{
-						selector: options?.selector,
-						key: 'textDecoration',
-						value: styleValue,
-					},
-			  ]
-			: [];
-	},
-};
-
-export default textDecoration;
